ENH: Implement action accelerator keys
Accelerator keys are a much quicker way to work with what kupfer excels
at: a choice of many actions on an object.

For example: The Open Terminal Here action has accelerator key T. While
a directory is selected in the first pane, pressing Alt + T will run the
open terminal here action directly.

If the action needs an indirect object, instead of running the action
the third pane will be selected immediately. For example the Open With
action has the accelerator key W, so you'd find a file, press Alt + W,
then type an application to open the file with.

Accelerators are configurable and you can add a custom one to every
action, including on actions in plugins, of course. You can also erase
or change the preset accelerator keys.

In this implementation, accelerators are always using the Alt modifier.