BUG: Fix “set as default for object” for text leaf and commands
TextLeaf and thus Command (shell command) used the hash() function to
make their repr key, which nowadays does not have a stable result across
saving and restoring. Use stable identifiers as repr_key so that text
leaves and commands can be used with “set (command) as default for
object” properly.