BUG: Use uri, not path to enqueue tracks in Rhythmbox
This fixes a path encoding issue: Gio.File.get_path raises an exception
if the path can not be decoded (for example if it is in a legacy
non-utf-8 encoding). Go around this issue by using the uri as it was
received from Rhythmbox, which is file system encoding agnostic.