Support file://<local-hostname>/ URLs when they are pasted
This change makes sure these links are recognized as files if they are
pasted into Kupfer.

ls --hyperlink produces URLs that look like this:

    file://<hostname>/path/to/kupfer/README.rst

where <hostname> is the hostname. Terminal will need to support
hyperlinks for this to be visible.

In the spirit of the hyperlink spec,
https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda
we should recognize URLs with the local hostname as local paths, and
those with mismatching hostname as not being local.